Bug fix to make duper work in larger tactic-style proofs
As best as I can tell, issues were arising in the barber_paradox_inline tests
because necessary information from the initial MetavarContext was not being
transferred through the whole pipeline. In particular, it was not available
when RuleM.isProof was called on 'b', which was problematic because the
Lean.Meta functions seem to use information from the MetavarContext. So I added
mctx to the state of ProverM.lean which appears to have resolved at least most
of the issue.

I'm not yet certain whether the changes I added were fully sufficient to make
duper actually work as part of larger tactic-style proofs in general (issues
with barber_paradox_inline3 may indicate that my fix was incomplete), but I
think it's an improvement on what we previously had.