Modified isMaximalLit and isMaximalInSubClause to support incomparable literals
The reason that some of the TPTP_tests were reaching saturation without obtaining
a contradiction was that there were clauses whose literals were incomparable. Because of
this, none of the literals in said clauses were identified as maximal, preventing
needed superposition inferences.

There is still one tptp test for which duper saturates prior to obtaining a contradiction,
but this appears to be due to a parsing bug in the tptp tactic, not due to a bug in the
duper tactic itself.