Editing superposition's side condition checks
Note: This commit seems to have either exposed or introduced some bugs which cause duper to perform
worse than in previous commits. I'm still pushing this so that I will be able to continue to work remotely,
but note that the previous commit can be regarded as the most recent working commit.