Changed Unif.lean to use Lean's built-in unifier
Additionally, also changed the types of SimpRule and BackwardSimpRule to
take Clauses rather than MClauses. This is beneficial for two reasons.

First, as noted in TODO.md, this gives simplification rules more freedom
about whether to call loadClause. There are some instances in which
calling loadClause may not be necessary in a simplification rule, and so
changing the implementation of that simplification rule may make it more
efficient.

Second, it would be difficult to make Match.lean use Lean's built-in
unifier without this change. The way Lean wants people to implement
matching is to use withNewMCtxDepth and isDefEq. In order to do this
though, it would be much more convenient for simplification rules to
receive Clauses rather than MClauses, so that each simplification rule
can determine whether a clause should be loaded inside withMCtxDepth or
not.