Updating Lean version to nightly-2022-11-19
Note, one of the changes that was made between the previous nightly
version and this commit's nightly version was several data structures
(e.g. HashMaps) were moved from the Std namespace to the Lean namespace.
But as a result, in this commit's nightly version, there is now a choice
between whether different versions of the same data structure (e.g.
there is both Std.HashMap and Lean.HashMap). For the purposes of
updating the Lean version, I chose to in general use the Lean
namespace's version of data structures (except for data structures like
Std.Data.BinomialHeap where the only version available was in Std). But
I don't know whether this choice was best or not. It may make sense to
change which versions of data structures Duper uses in a future commit.