Minor tweak to preprocessingClausification
This commit makes a small change to the order in which
preprocessingClausification adds clausified clauses to the passive set.
I'm not sure whether this change will be significant or not (and if it
is, whether it will improve duper's performance or make it worse), but
in the previous commit which introduced preprocessing clausification,
duper's performance became worse on average (though not strictly worse).
So I want to see whether tweaking with preprocessingClausification's
implementation can fix or impact that performance dip.