Superposition proof reconstruction bug fix
Fixed a bug in which Meta.kabstract (which was previously used in
simultaneous superposition's proof reconstruction code) abstracted at a
different set of positions than those that RuleM.replace replaced at.

This occurred because Meta.kabstract (a built-in Lean function) uses
definitional equality whereas RuleM.replace does not, and was
problematic because the clause output by superposition's main code was
different from the clause proven by the proof reconstruction code (at
least when RuleM.replace and Meta.kabstract disagreed on the set of
positions a subexpression appeared at in an expression). In the previous
commit (02071c28873ff11269d085f5be419bcc7ea2c767), this bug was
exhibited by FOF tests SYN396+1 and SYN962+1.