Removed unnecessary check in removeDescendants and fixed negative simplify reflect bug
Previously, in removeDescendants we checked that the direct descendant d
listed c (the clause whose descendants are being removed) as a
generating ancestor. The justification for this was that it was possible
for c to list d as a direct descendant but not have d list c as a
generating ancestor. This is indeed possible, but it's still okay to
remove d from the passive set because if it is needed, it can be
rederived from the clause that c has been simplified into.

The negative simplify reflect bug is entirely unrelated but yielded an
unsound proof in NUM925_1 in the previous commit.