Partial immediate simplification
After performing inference rules, but before adding the produced clauses
to the passive set, we check whether any of the produced clauses subsume
the given clause. If one of the produced clauses subsumes the given
clause, then we can remove the given clause and only add the clause that
subsumes it (as opposed to adding all of the clauses that were generated
by the inference rules).

This is not a full implementation of immediate simplification as
descrbied in
http://www.cs.man.ac.uk/~korovink/my_pub/iprover_ijcar_app_2020.pdf
because we do not currently inter-simplify the clauses that are produced
by each given clause.