Disabled orphan elimination
For some reason, the current implementation of orphan elimination seems
to make duper perform worse on average (though not strictly worse; there
are some tests that duper solves with orphan elimination that it does
not solve without). I've added as a TODO item for figuring out whether
orphan elimination simply isn't profitable for duper or whether there is
an issue with my current implementation.