Stopped using StrategyHeap
I'm still figuring out exactly why this is the case, but it seems that
locally, SYN007+1.014 only causes duper to crash if we use
StrategyHeap.lean to choose the givenClause. I'm pushing a version of
duper that doesn't use StrategyHeap.lean to see if it resolves the
github test crashes. Even if it does, I'll try to determine what exactly
is/was causing the issue in the StrategyHeap implementation.

As an additional note, the previous implementation of chooseGivenClause
(prior to the introduction of StrategyHeap.lean) was buggy in that it
did not correctly oscillate between the age and weight heaps
(specifically, the fairness counter would get updated even if the
age heap popped a clause that had already been removed from the
passive set, effectively skipping a cycle in which the age heap should
have been used). This bug has been fixed for the current commit.