ExistsHoist
The existsHoist rule introduces a fresh variable that is not unified
with anything in particular. In order to support proof reconstruction
for this, it was necessary to modify the RuleM Proof structure to
include a new field "newVarIndices". The idea behind this field is that
if a rule introduces fresh metavariables that will not be unified or
instantiated with anything (and therefore will be removed and replaced
with bound variables in neutralizeMClause), those metavariables' ids
can be passed into yieldClause so that the ProofReconstructor can
know which variables in the clause it is interacting with correspond to
fresh variables generated by the current rule.

I also added code for the forallHoist rule, but this rule is not enabled
in the current commit because I'm running into issues successfully unifying
with forall expressions (I'm presently uncertain whether the issue lies in
the unifier or in my forallHoist code).