Updating (but not enabling) existsHoist and forallHoist
Added updates to the existsHoist and forallHoist code to handle metavariables
that may appear in the body of the quantified statement. I also added tests
for existsHoist and forallHoist to test_regression.lean (the hoist rules are
only tested if identBoolHoist is disabled).

Notably, two of the forallTests run into proof reconstruction errors when
identBoolHoist is enabled. I assume this is due to some bug in identBoolHoist
but I haven't actually looked into it yet.