Added some guard checks to boolSimp and forallHoist
The purpose of the guard checks is to prevent these rules from inadvertently
treating forall statements (that aren't implications) as implications or,
in the case of forallHoist, treating implications as general forall statements
(forallHoist shouldn't be performed on implications).

Note that although this resolved one of the issues exhibited when boolSimp
is enabled (neg1 in test_boolSimp.lean) it did not resolve the other two example
issues. For this reason, I left boolSimp disabled. However, I think the way to
resolve the other two example issues is to change the fold functions so that
no rule can modify expressions which other expressions depend on (since in
neg2 and neg3 of test_boolSimp.lean, the substitution that boolSimp is performing
is one that we actually want boolSimp to perform).