More progress towards type inhabitation reasoning
The main hurdle that remains is getting proof reconstruction to work. I
think that (for the most part), the current code does a good job only
ending the proof search when a non-vacuous empty clause should be provable
(though there are still some bugs to be worked out).

However, although the current code is able to keep track of which types
were verified to be inhabited, there are still some issues to be
resolved in connecting that to the actual proof reconstruction.