Handing off type inhabitation reasoning proof reconstruction
I removed most of the code that didn't align with the described sorry
approach, but left some stuff in Util/Misc.lean because without the
removed changes, it doesn't hurt anything, and I wasn't sure whether it
would be helpful. If it turns out to not be needed/wanted, it should be
easy enough to revert findInstance to how it was in the previous commit.