Made implications' lhs/rhs valid Expr positions
Previously, because implications are represented with forallE expressions
in Lean, we did not take the lhs or rhs of an implication to be a valid
Expr pos. This didn't cause any concrete problems on first-order tests
because clausification will always eventually remove the implications,
but it may have created future issues for higher order problems that
contain implications under binders.

Prior to this commit, if identBoolHoist was disabled, then PUZ137_8
would saturate (because duper could not handle the implication that
would occur under a binder). After this commit, PUZ137_8 can be solved
regardless of whether identBoolHoist is enabled.