Added beta/eta reduction prior to kbo comparisons
Although I tried to minimize instances in which the same expression
was repeatedly beta/eta reduced, I expect this commit to introduce
overhead and therefore negatively impact efficiency. I had to comment
out MGT008 in test_regression.lean because duper took too long to solve
it after this change.

Depending on how widespread/significant the overhead is (as revealed by
the github tests) it may be worthwhile to explore ways to further
reduce this commit's overhead (though some amount is
inevitable/necessary).