Enabled fluidSup
fluidSup does appear to be working (and when it is applied, proof
reconstruction appears to be successful). However, it does not appear
to be successfully applied in all of the places I would expect it to
be (for an example, see supWithLambdaEx13 in fluidSupTests.lean). I
am still investigating with whether this is due to an issue in how I
set up the unification problem in fluidSup.lean or whether this is an
issue where the unification generator is not generating the necessary
unifier.