Added but didn't enable fluidBoolHoist
I've been unable to properly test fluidBoolHoist because enabling it
exposes what I believe to be an unrelated bug.

The bug can be seen in ex27 of test_higher_order.lean (when fluidBoolHoist
is enabled). The gist is that ex27 will use NeHoist to unify `?m e` with
`?m3 B =/= ?m4 B` (which is fine in principle and leads to a contradiction).
The issue is that in order to perform this unification, NeHoist unifies
`?m e` with `@Ne.{?lvlmvar} ?m1 ?m2` and the unification fails to instantiate
?lvlmvar with the level corresponding to the sort of the values ?m1 and ?m2
are instantiated with. As a result, in proof reconstruction, the parent clause
is instantiated to have the literal `@Ne.{?lvlmvar} (?m3 B) (?m4 B)` when it
should have `@Ne.{1} (?m3 B) (?m4 B)` (because `?m3 B` and `?m4 B` have Sort 1).
This discrepancy creates an issue that causes NeHoist's proof reconstruction to fail.