Simultaneous superposition bug fix
We need to call betaEtaReduce on mainPremise before calling
replaceGreenWithPos to ensure that equivalent expressions are actually
replaced. For instance, it is possible for `f` to be unified with
`(fun a => f a)`, but unless if we betaEtaReduce the latter expression,
replaceGreenWithPos will fail to recognize the two as equal (and
therefore perform the appropriate replacement).