Changed how potentiallyVacuousClauses are handled
The previous strategy for handling potentially vacuous clauses was
deficient in that it could prevent duper from making necessary
inferences by refusing to reason about clauses with bvars that don't
appear in the body and have potentially empty types.

The reason this was problematic is that if we had a clause like
(x : t1) -> True = False v Nonempty t2 = True, refusing to reason about
this clause could block Duper from deriving that (t1 -> t2) is Nonempty.

This commit updates the strategy for defining and handling potentially
vacuous clauses to avoid this issue. Now, even if a clause has bvars
that don't appear in the clause body and have potentially empty types,
Duper will still reason about the clause (performing forward
simplification rules and inference rules on it). However, if a clause is
potentially vacuous (meaning it has a bvar type that may be empty),
duper will not use the clause for any backward simplification rules, as
this could cause duper to replace a legitimately informative clause with
a vacuously true clause.

Upon learning that a new type is nonempty, Duper will iterate through
the set of clauses marked potentially vacuous and re-evaluate whether
any additional work needs to be done on them (e.g. making them eligible
for backward simplification rules or removing a vanished bvar).

Athough this commit improves how potentiallyVacuousClauses are defined
and handled, there is still work to be done on the inhabitation
reasoning front specifically concerning modifying deriveNonempty and
Lean.Meta.findInstance to enable deriving inhabitation facts from
multiple clauses (or from a combination of clauses, typeclass inference,
and local context witnesses).