Working on improving sequential quantifier skolemization
The current way we clausify something like "(Exists x : t1, Exists y :
t2, p x y) = True" is to first generate a clause where x is skolemized,
"(Exists y : t2, p skS0 y) = True", and then generate a clause where y is
skolemized afterwards "p skS0 skS1 = True". The issue with this is that
skolem symbols are functions that take in as arguments all of the
clause's bvars. A consequence of this approach is that while skS0 will
have the type "[clause bvars] -> t1 -> t1" (as it should), skS1 will
have the type "[clause bvars] -> t1 -> t2 -> t2". The issue is that
because skS1 is skolemized in a clause where skS0 was already
skolemized, t1 becomes one of the bvarTypes for the clause where skS1 is
skolemized, making skS1 depend on the provided value for t1 that should
only need to impact skS0. When unnecessary dependencies of this form
accumulate, it can make unification problems that ought to be solvable
impossible, preventing duper from making potentially necessary
inferences.

To attempt to resolve this, this commit works towards an alternative
approach of clausifying clauses of the form "(Exists x, Exists y, ...
Exists z, b) = True" and clauses of the form "(Forall x, Forall y, ...
Forall z, b) = False". The strategy is to, when sequential existential
quantifiers (equated to True) or sequential universal quantifiers
(equated to False) are clausified, rather than skolemizing the
quantifiers one at a time, we transform the sequence of quantifiers into
one single quantifier of the product of the types being quantified. So,
for instance, "Exists x : t1, Exists y : t2, p x y" would become
"Exists z : t1 x t2, p z.1 z.2". The hope is that this would ensure that
when the skolemization finally does occur, the generated skolem symbols
have no unnecessary dependencies.

Currently, although the code to clausify in the strategy described above
is written, it is disabled (via the option simultaneousSkolemization
which is set to false by default) because there are still some kinks to
work out. Specifically, when the option is enabled, the created clauses
interact poorly with simultaneousSuperposition because
Expr.replaceGreenWithPos currently does not recognize "x.fst" and "x.1"
as the same (this causes the organization example in
test_regression.lean to fail when both simultaneousSkolemization and
simultaneousSuperposition are enabled). Additionally, there appear to
be other currently undiagnosed issues such as the fact that SEU139 in
test_regression.lean will fail if simultaneousSkolemization is enabled.

My plan is to have simultaneousSkolemization enabled by default once I
can resolve these bugs and issues, because I think that generally
speaking, it is a better strategy than sequential skolemization.