Improving portfolio mode's spread of options
This commit adds more duper instances for the portfolio mode to possibly
choose from. It will still only choose 4 or 5 at a time, but which 4 or 5 can
depend on whether the user specifies the inhabitationReasoning option
and monomorphization option.

Instance 0 is now treated as a special instance that allows the user to
more manually select their options (inhabitationReasoning,
monomorphization, includeExpensiveRules, and selFunction are the
currently available options). All other instances have all options
specified, reducing the size of the suggestion returned by `duper?`.