Adding distanceFromGoal clause selection penalty
In this commit, the penalty does not do very much because there isn't a
good way to track which clauses originally came from the goal through
the monomorphization procedure. However, once that functionality is
added, hopefully Duper will be better able to handle the presence of
irrelevant lemmas (though of course, testing will be necessary to
confirm whether the distanceFromGoal penalty is actually net
beneficial).